requirements.txt
## Dependencias Jupyter
tensorflow>=2.13.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
ipykernel>=6.25.0
jupyter>=1.0.0

## Dependencias API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
pydantic>=2.4.0
python-multipart>=0.0.6
httpx>=0.25.0
requests>=2.31.0
